Update program 07.py

+++ ProgramBatch1/program 07.py
@@ -1,2 +1,2 @@
-nums = [int(input("Enter a number: ")) for _ in range (10)]
+nums = [float(input("Enter a number: ")) for _ in range (10)]
 print ("Sum:", sum(nums))
